% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary
Provides a brief summary of the package libraries on your machine.
Tamas!}
\usage{
lib_summary()
}
\value{
A data.frame containing the count of packages in each othe libraries
}
\description{
R Library Summary
Provides a brief summary of the package libraries on your machine.
Tamas!
}
