% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylibminer-package.R
\docType{package}
\name{mylibminer-package}
\alias{mylibminer}
\alias{mylibminer-package}
\title{mylibminer: Explore your R libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pongi09/mylibminer}
  \item Report bugs at \url{https://github.com/pongi09/mylibminer/issues}
}

}
\author{
\strong{Maintainer}: Tamas Pongracz \email{dsdpongi@gmail.com} (\href{https://orcid.org/XXXX-XXXX-XXXX-XXXX}{ORCID})

}
\keyword{internal}
